IndexedDB: Fix trivial cpplint.py errors
Classes of errors tackled: whitespace, include-what-you-use, explicit
constructors, parameter names, include guards, bracing, and one
use of "unsigned long" in IPC.

This does not tackle use of non-const references, since those should
be carefully reviewed and affects the callers. Also, several uses
of "long" remain in content/child when implementing Blink interfaces.

R=cmumford@chromium.org,ericu@chromium.org,tsepez@chromium.org
BUG=381456

Review URL: https://codereview.chromium.org/311263014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276451 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4cfd02d3569fcdd2066b41b382e377065485d065

+++ content/browser/indexed_db/indexed_db_dispatcher_host.cc
@@ -244,7 +244,6 @@ IndexedDBCursor* IndexedDBDispatcherHost::GetCursorFromId(int32 ipc_cursor_id) {
            web_metadata.object_stores.begin();
        iter != web_metadata.object_stores.end();
        ++iter) {
-
     const content::IndexedDBObjectStoreMetadata& web_store_metadata =
         iter->second;
     ::IndexedDBObjectStoreMetadata idb_store_metadata;
@@ -902,7 +901,7 @@ void IndexedDBDispatcherHost::CursorDispatcherHost::OnAdvance(
     int32 ipc_cursor_id,
     int32 ipc_thread_id,
     int32 ipc_callbacks_id,
-    unsigned long count) {
+    uint32 count) {
   DCHECK(
       parent_->indexed_db_context_->TaskRunner()->RunsTasksOnCurrentThread());
   IndexedDBCursor* idb_cursor =
